Add custom data sources (#1713)
* Refactor Batch datasources into relevant offline store files


* Move batch data sources to corresponding offline stores


* change default values to None


* implement from_proto for File and Bigquery, and move logic out of data_source.py


* Fix lint


* Fix lint after rebase


* more refactoring


* remove prints from test


* make format


* Remove bad assert


* move redshift over as well


* Remove print


* Remove incorrect logic from the get_data_source_class_from_type


* fix import


* make format
