+++ sdk/python/feast/repo_config.py
@@ -195,6 +195,11 @@ class FeastConfigError(Exception):
         )
 
 
+def get_data_source_class_from_type(data_source_type: str):
+    module_name, config_class_name = data_source_type.rsplit(".", 1)
+    return get_class_from_type(module_name, config_class_name, "Source")
+
+
 def get_online_config_from_type(online_store_type: str):
     if online_store_type in ONLINE_STORE_CLASS_FOR_TYPE:
         online_store_type = ONLINE_STORE_CLASS_FOR_TYPE[online_store_type]
